finished 8 section

+++ src/controllers/videoController.js
@@ -22,32 +22,44 @@ export const home = async(req, res) => {
 };
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
-    console.log(video);     //it keeps showing new ObjectId(...) 
-    
-    const owner = await User.findById(video.owner);
+    const video = await Video.findById(id).populate("owner");
+    console.log(video);
+    //const owner = await User.findById(video.owner.toHexString());
+        //it keeps showing new ObjectId(...) 
     if (!video) {
         return res.render("404", {pageTitle: "Video not found."} ); 
     }
+   
     return res.render("watch", { pageTitle: video.title, video});
 };
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session;
     const video = await Video.findById(id);
     if (!video) {
         return res.render("404", {pageTitle: "Video not found."} ); 
     }
+    console.log(video.owner, _id);
+    //we should protect edit page on back-end as well (only owners can edit their videos)
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+    
     return res.render("edit", { pageTitle: `Edit ${video.title}`, video});
 };
 
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session;
     const {title, description, hashtags} = req.body;
     const video = await Video.exists({_id: id});
     console.log(req.body);  
     if (!video) {
         return res.status(404).render("404", {pageTitle: "Video not found."} ); 
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -67,16 +79,18 @@ export const postUpload = async(req, res) => {
     } = req.session;
     const {path: fileUrl}= req.file;        // power of ES6
     const {title, description, hashtags} = req.body;
-    console.log(req.body);  
+    //console.log(req.body);  
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
-            fileUrl,
-            owner: _id,
+            fileUrl, 
+            owner:_id,
             hashtags: Video.formatHashtags(hashtags),
-            
         }); 
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
     // waiting for video to be saved in the db 
     //await video.save();
     // here we will add a video to the videos array
@@ -92,6 +106,15 @@ export const postUpload = async(req, res) => {
 export const deleteVideo = async(req,res) => {
     const { id } = req.params;
     console.log(id);
+    const { user: {_id} } = req.session;
+    //we don't need populate function because we don't need whole user
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render("404", {pageTitle: "Video not found."} ); 
+    }
+    if (String(video.owner) !== String(_id)) {
+            return res.status(403).redirect("/");
+    }
     //delete video
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
